api_leader_v2/requirements.txt
Django==4.2.8
psycopg2-binary==2.9.9
djangorestframework==3.14.0
python-dotenv==1.0.1
ipdb==0.13.13
